mv visualize.py -> app.py; restore deprecated/visualize.py
