+++ openadapt/visualize.py
@@ -1,151 +1,214 @@
 """Implements visualization utilities for OpenAdapt."""
 
-from base64 import b64encode
-from functools import partial
-from os import path, sep
 from pprint import pformat
+from threading import Timer
+import html
+import os
+import string
 
+from bokeh.io import output_file, show
+from bokeh.layouts import layout, row
+from bokeh.models.widgets import Div
 from loguru import logger
-from nicegui import events, ui
-from notifypy import Notify
 from tqdm import tqdm
-import click
+import fire
 
-from openadapt import config
-from openadapt.db.crud import get_latest_recording, get_recording
+from openadapt import config, video
+from openadapt.db.crud import get_latest_recording
 from openadapt.events import get_events
+from openadapt.models import Recording
+from openadapt.privacy.providers.presidio import PresidioScrubbingProvider
 from openadapt.utils import (
     EMPTY,
+    compute_diff,
     configure_logging,
     display_event,
+    evenly_spaced,
     image2utf8,
-    plot_performance,
     row2dict,
     rows2dicts,
 )
 
 SCRUB = config.SCRUB_ENABLED
-
 if SCRUB:
-    # too many warnings from scrubbing
-    __import__("warnings").filterwarnings("ignore", category=DeprecationWarning)
-    from openadapt.privacy.providers.presidio import PresidioScrubbingProvider
-
     scrub = PresidioScrubbingProvider()
-
 
 LOG_LEVEL = "INFO"
 MAX_EVENTS = None
+MAX_TABLE_CHILDREN = 5
+MAX_TABLE_STR_LEN = 1024
 PROCESS_EVENTS = True
+IMG_WIDTH_PCT = 60
+CSS = string.Template("""
+    table {
+        outline: 1px solid black;
+    }
+    table th {
+        vertical-align: top;
+    }
+    .screenshot img {
+        display: none;
+        width: ${IMG_WIDTH_PCT}vw;
+    }
+    .screenshot img:nth-child(1) {
+        display: block;
+    }
 
-performance_plot_img: ui.interactive_image = None
+    .screenshot:hover img:nth-child(1) {
+        display: none;
+    }
+    .screenshot:hover img:nth-child(2) {
+        display: block;
+    }
+    .screenshot:hover img:nth-child(3) {
+        display: none;
+    }
+
+    .screenshot:active img:nth-child(1) {
+        display: none;
+    }
+    .screenshot:active img:nth-child(2) {
+        display: none;
+    }
+    .screenshot:active img:nth-child(3) {
+        display: block;
+    }
+""").substitute(
+    IMG_WIDTH_PCT=IMG_WIDTH_PCT,
+)
 
 
-def create_tree(
-    tree_dict: dict, max_children: str = config.VISUALIZE_MAX_TABLE_CHILDREN
-) -> list[dict]:
-    """Recursively creates a tree from a dictionary.
+def recursive_len(lst: list, key: str) -> int:
+    """Calculate the recursive length of a list based on a key.
 
     Args:
-        tree_dict (dict): The dictionary to create a tree from.
-        max_children (str, optional): The maximum number of children to display.
-        Defaults to MAX_TABLE_CHILDREN.
+        lst (list): The list to calculate the length of.
+        key: The key to access the sublists.
 
     Returns:
-        list[dict]: Data for a Quasar Tree.
+        int: The recursive length of the list.
     """
-    tree_data = []
-    for key, value in tree_dict.items():
-        if value in EMPTY:
+    try:
+        _len = len(lst)
+    except TypeError:
+        return 0
+    for obj in lst:
+        try:
+            _len += recursive_len(obj.get(key), key)
+        except AttributeError:
             continue
-        node = {
-            "id": (
-                str(key)
-                + f"{(': '  + str(value)) if not isinstance(value, (dict, list)) else ''}"  # noqa
-            )
-        }
-        if isinstance(value, dict):
-            node["children"] = create_tree(value)
-        elif isinstance(value, list):
-            if max_children is not None and len(value) > max_children:
-                node["children"] = create_tree(
-                    {i: v for i, v in enumerate(value[:max_children])}
-                )
-                node["children"].append({"id": "..."})
-            else:
-                node["children"] = create_tree({i: v for i, v in enumerate(value)})
-        tree_data.append(node)
-    return tree_data
+    return _len
 
 
-def set_tree_props(tree: ui.tree) -> None:
-    """Sets properties for a UI tree based on values from config.
+def format_key(key: str, value: list) -> str:
+    """Format a key and value for display.
 
     Args:
-      tree (ui.tree): A Quasar Tree.
+        key: The key to format.
+        value: The value associated with the key.
+
+    Returns:
+        str: The formatted key and value.
     """
-    tree._props["dense"] = config.VISUALIZE_DENSE_TREES
-    tree._props["no-transition"] = not config.VISUALIZE_ANIMATIONS
-    tree._props["default-expand-all"] = config.VISUALIZE_EXPAND_ALL
-    tree._props["filter"] = ""
+    if isinstance(value, list):
+        return f"{key} ({len(value)}; {recursive_len(value, key)})"
+    else:
+        return key
 
 
-def set_filter(
-    text: str,
-    tree: ui.tree,
-) -> None:
-    """Sets filter for UI trees.
+def indicate_missing(some: list, every: list, indicator: str) -> list:
+    """Indicate missing elements in a list.
 
     Args:
-        tree (ui.tree): A Quasar Tree.
-        text (str): The text to filter by.
+        some (list): The list with potentially missing elements.
+        every (list): The reference list with all elements.
+        indicator (str): The indicator to use for missing elements.
+
+    Returns:
+        list: The list with indicators for missing elements.
     """
-    tree._props["filter"] = text
-    tree.update()
+    rval = []
+    some_idx = 0
+    every_idx = 0
+    while some_idx < len(some):
+        skipped = False
+        while some[some_idx] != every[every_idx]:
+            every_idx += 1
+            skipped = True
+        if skipped:
+            rval.append(indicator)
+        rval.append(some[some_idx])
+        some_idx += 1
+        every_idx += 1
+    return rval
 
 
-def toggle_dark_mode(
-    ui_dark: ui.dark_mode, plots: tuple[str], curr_logo: ui.image, images: tuple[str]
-) -> None:
-    """Handles dark mode toggle.
+def dict2html(
+    obj: dict,
+    max_children: int = MAX_TABLE_CHILDREN,
+    max_len: int = MAX_TABLE_STR_LEN,
+) -> str:
+    """Convert a dictionary to an HTML representation.
 
     Args:
-        ui_dark (ui.dark_mode): The dark mode switch.
-        plots (tuple[str]): The performance plots.
-        curr_logo (ui.image): The current logo.
-        images (tuple[str]): The light and dark mode logos (decoded)
+        obj (dict): The dictionary to convert.
+        max_children (int): The maximum number of child elements to display in a table.
+        max_len (int): The maximum length of a string value in the HTML representation.
+
+    Returns:
+        str: The HTML representation of the dictionary.
     """
-    global performance_plot_img
-    ui_dark.toggle()
-    ui_dark.update()
-    config.persist_env("VISUALIZE_DARK_MODE", ui_dark.value)
-    curr_logo.source = images[int(ui_dark.value)]
-    curr_logo.update()
-    performance_plot_img.source = plots[int(ui_dark.value)]
-    performance_plot_img.update()
+    if isinstance(obj, list):
+        children = [dict2html(value, max_children) for value in obj]
+        if max_children is not None and len(children) > max_children:
+            all_children = children
+            children = evenly_spaced(children, max_children)
+            children = indicate_missing(children, all_children, "...")
+        html_str = "\n".join(children)
+    elif isinstance(obj, dict):
+        rows_html = "\n".join([f"""
+                <tr>
+                    <th>{format_key(key, value)}</th>
+                    <td>{dict2html(value, max_children)}</td>
+                </tr>
+            """ for key, value in obj.items() if value not in EMPTY])
+        html_str = f"<table>{rows_html}</table>"
+    else:
+        html_str = html.escape(str(obj))
+        if len(html_str) > max_len:
+            n = max_len // 2
+            head = html_str[:n]
+            tail = html_str[-n:]
+            snipped = html_str[n:-n]
+            middle = f"<br/>...<i>(snipped {len(snipped):,})...</i><br/>"
+            html_str = head + middle + tail
+    return html_str
 
 
 @logger.catch
-@click.command()
-@click.option(
-    "--timestamp",
-    type=str,
-    help="The timestamp of the recording to visualize.",
-)
-def main(timestamp: str, notify: bool = True) -> None:
-    """Visualize a recording."""
+def main(
+    recording: Recording = None,
+    diff_video: bool = False,
+    cleanup: bool = True,
+) -> bool:
+    """Visualize a recording.
+
+    Args:
+        recording (Recording, optional): The recording to visualize.
+        diff_video (bool): Whether to diff Screenshots against video frames.
+        cleanup (bool): Whether to remove the HTML file after it is displayed.
+
+    Returns:
+        bool: True if visualization was successful, None otherwise.
+    """
     configure_logging(logger, LOG_LEVEL)
 
-    ui_dark = ui.dark_mode(config.VISUALIZE_DARK_MODE)
-
-    if timestamp is None:
+    if recording is None:
         recording = get_latest_recording()
-    else:
-        recording = get_recording(timestamp)
-
     if SCRUB:
         scrub.scrub_text(recording.task_description)
-    logger.debug(f"{recording=}")
+    logger.info(f"{recording=}")
+    logger.info(f"{diff_video=}")
 
     meta = {}
     action_events = get_events(recording, process=PROCESS_EVENTS, meta=meta)
@@ -156,142 +219,45 @@ def main(timestamp: str, notify: bool = True) -> None:
     logger.info(f"event_dicts=\n{pformat(event_dicts)}")
 
     recording_dict = row2dict(recording)
-
     if SCRUB:
         recording_dict = scrub.scrub_dict(recording_dict)
 
-    # setup tables for recording / metadata
-    recording_column = [
-        (
-            {
-                "name": key,
-                "field": key,
-                "label": key,
-                "sortable": False,
-                "required": False,
-                "align": "left",
-            }
-        )
-        for key in recording_dict.keys()
-    ]
-
-    meta_col = [
-        {
-            "name": key,
-            "field": key,
-            "label": key,
-            "sortable": False,
-            "required": False,
-            "align": "left",
-        }
-        for key in meta.keys()
-    ]
-
-    plots = (
-        plot_performance(recording.timestamp, save_file=False, view_file=False),
-        plot_performance(
-            recording.timestamp, save_file=False, view_file=False, dark_mode=True
+    rows = [
+        row(
+            Div(
+                text=f"<style>{CSS}</style>",
+            ),
         ),
-    )
-
-    with ui.row():
-        with ui.avatar(color="auto", size=128):
-            images = ()
-
-            # generate base64 encoded images for light and dark mode
-            for i in range(2):
-                fp = f"{path.dirname(__file__)}{sep}app{sep}assets{sep}logo"
-                logo_base64 = b64encode(
-                    open(
-                        f"{fp}{'_inverted' if i > 0 else ''}.png",
-                        "rb",
-                    ).read()
-                )
-                images += (
-                    bytes(
-                        f"data:image/png;base64,{(logo_base64.decode('utf-8'))}",
-                        encoding="utf-8",
-                    ).decode("utf-8"),
-                )
-            curr_logo = ui.image(images[int(ui_dark.value)])
-        ui.switch(
-            text="Dark Mode",
-            value=ui_dark.value,
-            on_change=partial(toggle_dark_mode, ui_dark, plots, curr_logo, images),
-        )
-
-    # create splitter with recording info on left and performance plot on right
-    with ui.splitter(value=20).style("flex-wrap: nowrap;") as splitter:
-        splitter._props["limits"] = [20, 80]
-
-        # TODO: find a way to set "overflow: hidden;" for the splitter
-        with splitter.before:
-            ui.table(rows=[meta], columns=meta_col).style("min-width: 50em;")._props[
-                "grid"
-            ] = True
-        with splitter.after:
-            global performance_plot_img
-            performance_plot_img = ui.interactive_image(
-                source=plots[int(ui_dark.value)]
-            )
-            with performance_plot_img:
-                # ui.button(
-                #     on_click=lambda: plot_performance(
-                #         recording.timestamp, view_file=True, save_file=False
-                #     ),
-                #     icon="visibility",
-                # ).props("flat fab").tooltip("View")
-
-                ui.button(
-                    on_click=lambda: plot_performance(
-                        recording.timestamp,
-                        save_file=True,
-                        view_file=False,
-                        dark_mode=ui_dark.value,
-                    ),
-                    icon="save",
-                ).props("flat fab").tooltip("Save as PNG")
-
-    ui.table(rows=[recording_dict], columns=recording_column)
-
-    interactive_images = []
-    action_event_trees = []
-    window_event_trees = []
-    text_inputs = []
-
+        row(
+            Div(
+                text=f"{dict2html(recording_dict)}",
+            ),
+        ),
+        row(
+            Div(
+                text=f"{dict2html(meta)}",
+                width_policy="max",
+            ),
+        ),
+    ]
     logger.info(f"{len(action_events)=}")
+
+    if diff_video:
+        video_file_name = video.get_video_file_name(recording.timestamp)
+        timestamps = [
+            action_event.screenshot.timestamp - recording.video_start_time
+            for action_event in action_events
+        ]
+        frames = video.extract_frames(video_file_name, timestamps)
 
     num_events = (
         min(MAX_EVENTS, len(action_events))
         if MAX_EVENTS is not None
         else len(action_events)
     )
-
-    # global search
-    def on_change_closure(e: events.ValueChangeEventArguments) -> None:
-        for tree in range(len(action_event_trees)):
-            set_filter(
-                e.value,
-                action_event_trees[tree],
-            )
-            set_filter(
-                e.value,
-                window_event_trees[tree],
-            )
-
-    text_inputs.append(
-        ui.input(
-            label="search all",
-            placeholder="filter all",
-            on_change=partial(on_change_closure),
-        ).tooltip(
-            "this will search all trees, but can be overridden by individual filters"
-        )
-    )
-
     with tqdm(
         total=num_events,
-        desc="Generating Visualization" if not SCRUB else "Scrubbing Visualization",
+        desc="Preparing HTML",
         unit="event",
         colour="green",
         dynamic_ncols=True,
@@ -300,18 +266,32 @@ def main(timestamp: str, notify: bool = True) -> None:
             if idx == MAX_EVENTS:
                 break
 
-            image = display_event(action_event)
-            # diff = display_event(action_event, diff=True)
-            # mask = action_event.screenshot.diff_mask
+            try:
+                image = display_event(action_event)
+            except TypeError as exc:
+                # https://github.com/moses-palmer/pynput/issues/481
+                logger.warning(exc)
+                continue
+
+            if diff_video:
+                frame_image = frames[idx]
+                diff_image = compute_diff(frame_image, action_event.screenshot.image)
+
+                # TODO: rename
+                diff = frame_image
+                mask = diff_image
+            else:
+                diff = display_event(action_event, diff=True)
+                mask = action_event.screenshot.diff_mask
 
             if SCRUB:
                 image = scrub.scrub_image(image)
-            #    diff = scrub.scrub_image(diff)
-            #    mask = scrub.scrub_image(mask)
+                diff = scrub.scrub_image(diff)
+                mask = scrub.scrub_image(mask)
 
             image_utf8 = image2utf8(image)
-            # diff_utf8 = image2utf8(diff)
-            # mask_utf8 = image2utf8(mask)
+            diff_utf8 = image2utf8(diff)
+            mask_utf8 = image2utf8(mask)
             width, height = image.size
 
             action_event_dict = row2dict(action_event)
@@ -321,101 +301,69 @@ def main(timestamp: str, notify: bool = True) -> None:
                 action_event_dict = scrub.scrub_dict(action_event_dict)
                 window_event_dict = scrub.scrub_dict(window_event_dict)
 
-            with ui.column():
-                with ui.row():
-                    interactive_images.append(
-                        ui.interactive_image(
-                            source=image_utf8,
-                        ).classes("drop-shadow-md rounded")
-                    )
-            with ui.splitter(value=60) as splitter:
-                splitter.classes("w-full h-full")
-                with splitter.after:
-                    ui.label("action_event_dict").style("font-weight: bold;")
-
-                    def on_change_closure(
-                        e: events.ValueChangeEventArguments, idx: int
-                    ) -> None:
-                        return set_filter(
-                            e.value,
-                            action_event_trees[idx],
-                        )
-
-                    text_inputs.append(
-                        ui.input(
-                            label="search",
-                            placeholder="filter",
-                            on_change=partial(
-                                on_change_closure,
-                                idx=idx,
-                            ),
-                        )
-                    )
-                    ui.html("<br/>")
-                    action_event_tree = create_tree(action_event_dict)
-                    action_event_trees.append(
-                        ui.tree(
-                            action_event_tree,
-                            label_key="id",
-                            # on_select=lambda e: ui.notify(e.value),
-                        )
-                    )
-                    set_tree_props(action_event_trees[idx])
-                with splitter.before:
-                    ui.label("window_event_dict").style("font-weight: bold;")
-
-                    def on_change_closure(
-                        e: events.ValueChangeEventArguments, idx: int
-                    ) -> None:
-                        return set_filter(
-                            e.value,
-                            window_event_trees[idx],
-                        )
-
-                    text_inputs.append(
-                        ui.input(
-                            label="search",
-                            placeholder="filter",
-                            on_change=partial(
-                                on_change_closure,
-                                idx=idx,
-                            ),
-                        )
-                    )
-                    ui.html("<br/>")
-                    window_event_tree = create_tree(window_event_dict, None)
-
-                    window_event_trees.append(
-                        ui.tree(
-                            window_event_tree,
-                            label_key="id",
-                            # on_select=lambda e: ui.notify(e.value),
-                        )
-                    )
-
-                    set_tree_props(window_event_trees[idx])
+            rows.append(
+                [
+                    row(
+                        Div(
+                            text=f"""
+                            <div class="screenshot">
+                                <img
+                                    src="{image_utf8}"
+                                    style="
+                                        aspect-ratio: {width}/{height};
+                                    "
+                                >
+                                <img
+                                    src="{diff_utf8}"
+                                    style="
+                                        aspect-ratio: {width}/{height};
+                                    "
+                                >
+                                <img
+                                    src="{mask_utf8}"
+                                    style="
+                                        aspect-ratio: {width}/{height};
+                                    "
+                                >
+                            </div>
+                            <table>
+                                {dict2html(window_event_dict , None)}
+                            </table>
+                        """,
+                        ),
+                        Div(text=f"""
+                            <table>
+                                {dict2html(action_event_dict)}
+                            </table>
+                        """),
+                    ),
+                ]
+            )
 
             progress.update()
 
         progress.close()
 
-    if notify:
-        Notify("Status", "Visualization finished", "OpenAdapt").send()
-        if not config.VISUALIZE_RUN_NATIVELY:
-            Notify(
-                "Status",
-                "Press Ctrl+C in the terminal to stop the visualization",
-                "OpenAdapt",
-            ).send()
+    title = f"recording-{recording.id}"
+    fname_out = f"recording-{recording.id}.html"
+    logger.info(f"{fname_out=}")
+    output_file(fname_out, title=title)
 
-    ui.run(
-        reload=False,
-        title=f"OpenAdapt: recording-{recording.id}",
-        favicon="📊",
-        native=config.VISUALIZE_RUN_NATIVELY,
-        fullscreen=config.VISUALIZE_RUN_NATIVELY,
+    result = show(  # noqa: F841
+        layout(
+            rows,
+        )
     )
+
+    def _cleanup() -> None:
+        os.remove(fname_out)
+        removed = not os.path.exists(fname_out)
+        logger.info(f"{removed=}")
+
+    if cleanup:
+        Timer(1, _cleanup).start()
+    return True
 
 
 if __name__ == "__main__":
-    main()
+    fire.Fire(main)
